+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add: add option to provide additional error context in `pushError` api.
 - Fix: Use `globalThis` instead of `global` or `window` in case the SDK is used in webworkers.
+- Add: Add batch execution support for transports in core.
 
 ## 1.0.3 – 1.0.5